AppleDiseaseScanner/model.py
import os
import logging
import numpy as np
from PIL import Image
import joblib
from sklearn.ensemble import RandomForestClassifier
from sklearn.preprocessing import StandardScaler

# Set up logging
logging.basicConfig(level=logging.DEBUG)

# Define the class labels
CLASS_LABELS = ['apple_scab', 'black_rot', 'cedar_apple_rust', 'healthy']

def load_model():
    """Load or create a simple model for apple disease detection"""
    try:
        logging.info("Setting up model for apple disease detection...")
        
        # In a real application, you would load a pre-trained model here
        # For demo purposes, we'll create a simple random forest classifier
        model = RandomForestClassifier(n_estimators=50, random_state=42)
        scaler = StandardScaler()
        
        # Since we don't have actual training data, we'll pre-define some basic weights
        # In production, you would load your trained model from a file
        
        logging.info("Model initialized successfully")
        return {'model': model, 'scaler': scaler}
    
    except Exception as e:
        logging.error(f"Error setting up model: {str(e)}")
        raise

def extract_features(image):
    """Extract simple features from the image"""
    # Convert to RGB and resize
    image = image.convert('RGB')
    image = image.resize((100, 100))
    
    # Convert to numpy array
    img_array = np.array(image)
    
    # Simple feature extraction: color statistics from each channel
    features = []
    
    # Mean and std of each color channel
    for i in range(3):  # RGB channels
        channel = img_array[:, :, i].flatten()
        features.extend([
            np.mean(channel),
            np.std(channel),
            np.percentile(channel, 25),
            np.percentile(channel, 75)
        ])
    
    # Add some texture features (simple edge detection)
    for i in range(3):
        channel = img_array[:, :, i]
        dx = np.diff(channel, axis=0)
        dy = np.diff(channel, axis=1)
        features.extend([
            np.mean(np.abs(dx)),
            np.mean(np.abs(dy))
        ])
    
    return np.array(features).reshape(1, -1)

def preprocess_image(image):
    """Preprocess the image for feature extraction"""
    try:
        # Extract features from the image
        features = extract_features(image)
        
        return features
    
    except Exception as e:
        logging.error(f"Error preprocessing image: {str(e)}")
        raise

def predict_disease(model_package, preprocessed_features):
    """Make a prediction on the preprocessed features using the model"""
    try:
        # For demonstration, we'll simulate classification
        # In a real app, you would use model_package['model'].predict(preprocessed_features)
        
        # Since we don't have actual trained weights, we'll simulate predictions based on features
        # This is just for demonstration and should be replaced with a real model in production
        features = preprocessed_features[0]
        
        # Simulate decision based on the image features
        # In a real application, you would use your trained model here
        r_mean, g_mean, b_mean = features[0], features[4], features[8]
        r_std, g_std, b_std = features[1], features[5], features[9]
        
        # Rules to simulate classification based on color patterns common in apple diseases
        # Note: This is an oversimplified simulation and NOT a real classifier
        
        # Very red or very green
        if r_mean > 150 and g_mean < 100:  # Reddish
            predicted_class_index = 1  # black_rot
            confidence = 75 + (r_mean - 150) / 3
        elif g_mean > 140 and r_mean < 120:  # Very green
            predicted_class_index = 3  # healthy
            confidence = 80 + (g_mean - 130) / 3
        # Yellow-orange tint (rust)
        elif r_mean > 120 and g_mean > 100 and b_mean < 80:
            predicted_class_index = 2  # cedar_apple_rust
            confidence = 70 + (r_mean + g_mean - b_mean) / 10
        # Dark spots (scab)
        elif r_std > 40 or g_std > 40:
            predicted_class_index = 0  # apple_scab
            confidence = 65 + r_std / 2
        else:
            # Default to healthy but with lower confidence
            predicted_class_index = 3  # healthy
            confidence = 60
        
        # Cap confidence at 95%
        confidence = min(confidence, 95)
        
        # Get the class label
        predicted_class = CLASS_LABELS[predicted_class_index]
        
        logging.info(f"Predicted class: {predicted_class}, Confidence: {confidence:.2f}%")
        
        return predicted_class, confidence
    
    except Exception as e:
        logging.error(f"Error during prediction: {str(e)}")
        raise
